batch dot WIP.
